SLI-1784 Adjust UI
